removed the general audio component and it's node editor, to use the media component and its node editor for audio instead
